requirements.txt
python-docx
markdown2
Pillow